Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
asyncpg
pydantic_settings
python-jose[cryptography]
passlib
pyyaml
